fix(#933): retain requested SRS name in CRS object for XML requests
When doing a GetFeature request to the WFS endpoint via a posted XML
query, the name of the SRS that was passed into the request to the user
was not retained, but an alias from the looked-up CRS was stated in the
returned features.
This only happened in XML POST requests, not in KVP GET requests.